+++ src/Console/ValidateSchemaCommand.php
@@ -6,10 +6,10 @@ use GraphQL\Type\Schema;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Cache\Repository as CacheRepository;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
-use Nuwave\Lighthouse\GraphQL;
 use Nuwave\Lighthouse\Schema\DirectiveLocator;
 use Nuwave\Lighthouse\Schema\Factories\DirectiveFactory;
 use Nuwave\Lighthouse\Schema\FallbackTypeNodeConverter;
+use Nuwave\Lighthouse\Schema\SchemaBuilder;
 use Nuwave\Lighthouse\Schema\TypeRegistry;
 
 class ValidateSchemaCommand extends Command
@@ -21,7 +21,7 @@ class ValidateSchemaCommand extends Command
     public function handle(
         CacheRepository $cache,
         ConfigRepository $config,
-        GraphQL $graphQL,
+        SchemaBuilder $schemaBuilder,
         DirectiveLocator $directiveLocator,
         TypeRegistry $typeRegistry
     ): void {
@@ -30,7 +30,7 @@ class ValidateSchemaCommand extends Command
             $config->get('lighthouse.cache.key')
         );
 
-        $originalSchema = $graphQL->prepSchema();
+        $originalSchema = $schemaBuilder->schema();
         $schemaConfig = $originalSchema->getConfig();
 
         // We add schema directive definitions only here, since it is very slow