+++ CHANGELOG.md
@@ -16,6 +16,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Changed
 
 - Optimize `@defer` by avoiding parsing the request multiple times
+- Move HTTP and Schema handling out of the GraphQL class https://github.com/nuwave/lighthouse/pull/1748
 
 ## 5.3.0